[Fetcher] Move internal classes to separate files
