+++ lib/bundler/fetcher.rb
@@ -6,6 +6,9 @@ module Bundler
 
   # Handles all the fetching with the rubygems server
   class Fetcher
+    require 'bundler/fetcher/dependency_fetcher'
+    require 'bundler/fetcher/index_fetcher'
+
     # This error is raised when it looks like the network is down
     class NetworkDownError < HTTPError; end
     # This error is raised if the API returns a 413 (only printed in verbose)
@@ -189,184 +192,6 @@ module Bundler
     end
 
   private
-
-    class FetcherImpl
-      attr_reader :connection
-      attr_reader :remote_uri
-      attr_reader :fetch_uri
-      attr_reader :display_uri
-
-      def initialize(connection, remote_uri, fetch_uri, display_uri)
-        raise 'Abstract class' if self.class == FetcherImpl
-        @connection = connection
-        @remote_uri = remote_uri
-        @fetch_uri = fetch_uri
-        @display_uri = display_uri
-      end
-
-      def api_available?
-        api_fetcher?
-      end
-
-      def api_fetcher?
-        false
-      end
-
-      def fetch(uri, counter = 0)
-        raise HTTPError, "Too many redirects" if counter >= Fetcher.redirect_limit
-
-        response = request(uri)
-        Bundler.ui.debug("HTTP #{response.code} #{response.message}")
-
-        case response
-        when Net::HTTPRedirection
-          new_uri = URI.parse(response["location"])
-          if new_uri.host == uri.host
-            new_uri.user = uri.user
-            new_uri.password = uri.password
-          end
-          fetch(new_uri, counter + 1)
-        when Net::HTTPSuccess
-          response.body
-        when Net::HTTPRequestEntityTooLarge
-          raise FallbackError, response.body
-        when Net::HTTPUnauthorized
-          raise AuthenticationRequiredError, remote_uri.host
-        else
-          raise HTTPError, "#{response.class}: #{response.body}"
-        end
-      end
-
-      def request(uri)
-        Bundler.ui.debug "HTTP GET #{uri}"
-        req = Net::HTTP::Get.new uri.request_uri
-        if uri.user
-          user = CGI.unescape(uri.user)
-          password = uri.password ? CGI.unescape(uri.password) : nil
-          req.basic_auth(user, password)
-        end
-        connection.request(uri, req)
-      rescue OpenSSL::SSL::SSLError
-        raise CertificateFailureError.new(uri)
-      rescue *HTTP_ERRORS => e
-        Bundler.ui.trace e
-        case e.message
-        when /host down:/, /getaddrinfo: nodename nor servname provided/
-          raise NetworkDownError, "Could not reach host #{uri.host}. Check your network " \
-          "connection and try again."
-        else
-          raise HTTPError, "Network error while fetching #{uri}"
-        end
-      end
-
-      def well_formed_dependency(name, *requirements)
-        Gem::Dependency.new(name, *requirements)
-      rescue ArgumentError => e
-        illformed = 'Ill-formed requirement ["#<YAML::Syck::DefaultKey'
-        raise e unless e.message.include?(illformed)
-        puts # we shouldn't print the error message on the "fetching info" status line
-        raise GemspecError,
-          "Unfortunately, the gem #{s[:name]} (#{s[:number]}) has an invalid " \
-          "gemspec. \nPlease ask the gem author to yank the bad version to fix " \
-          "this issue. For more information, see http://bit.ly/syck-defaultkey."
-      end
-    end
-
-    class DependencyFetcher < FetcherImpl
-      def api_available?
-        fetch(dependency_api_uri)
-      rescue NetworkDownError => e
-        raise HTTPError, e.message
-      rescue AuthenticationRequiredError
-        # We got a 401 from the server. Just fail.
-        raise
-      rescue HTTPError
-      end
-
-      def api_fetcher?
-        true
-      end
-
-      def specs(gem_names, full_dependency_list = [], last_spec_list = [])
-        query_list = gem_names - full_dependency_list
-
-        # only display the message on the first run
-        if Bundler.ui.debug?
-          Bundler.ui.debug "Query List: #{query_list.inspect}"
-        else
-          Bundler.ui.info ".", false
-        end
-
-        return {remote_uri => last_spec_list} if query_list.empty?
-
-        remote_specs = Bundler::Retry.new("dependency api", AUTH_ERRORS).attempts do
-          dependency_specs(query_list)
-        end
-
-        spec_list, deps_list = remote_specs
-        returned_gems = spec_list.map(&:first).uniq
-        specs(deps_list, full_dependency_list + returned_gems, spec_list + last_spec_list)
-      rescue HTTPError, MarshalError, GemspecError
-        Bundler.ui.info "" unless Bundler.ui.debug? # new line now that the dots are over
-        Bundler.ui.debug "could not fetch from the dependency API, trying the full index"
-        return nil
-      end
-
-      def dependency_specs(gem_names)
-        Bundler.ui.debug "Query Gemcutter Dependency Endpoint API: #{gem_names.join(',')}"
-        gem_list = []
-        deps_list = []
-
-        gem_names.each_slice(Source::Rubygems::API_REQUEST_SIZE) do |names|
-          marshalled_deps = fetch dependency_api_uri(names)
-          gem_list += Bundler.load_marshal(marshalled_deps)
-        end
-
-        spec_list = gem_list.map do |s|
-          dependencies = s[:dependencies].map do |name, requirement|
-            dep = well_formed_dependency(name, requirement.split(", "))
-            deps_list << dep.name
-            dep
-          end
-
-          [s[:name], Gem::Version.new(s[:number]), s[:platform], dependencies]
-        end
-
-        [spec_list, deps_list.uniq]
-      end
-
-      def dependency_api_uri(gem_names = [])
-        uri = fetch_uri + "api/v1/dependencies"
-        uri.query = "gems=#{URI.encode(gem_names.join(","))}" if gem_names.any?
-        uri
-      end
-    end
-
-    class IndexFetcher < FetcherImpl
-      def specs(_gem_names)
-        old_sources = Bundler.rubygems.sources
-        Bundler.rubygems.sources = [remote_uri.to_s]
-        Bundler.rubygems.fetch_all_remote_specs
-      rescue Gem::RemoteFetcher::FetchError, OpenSSL::SSL::SSLError => e
-        case e.message
-        when /certificate verify failed/
-          raise CertificateFailureError.new(display_uri)
-        when /401/
-          raise AuthenticationRequiredError, remote_uri
-        when /403/
-          if remote_uri.userinfo
-            raise BadAuthenticationError, remote_uri
-          else
-            raise AuthenticationRequiredError, remote_uri
-          end
-        else
-          Bundler.ui.trace e
-          raise HTTPError, "Could not fetch specs from #{display_uri}"
-        end
-      ensure
-        Bundler.rubygems.sources = old_sources
-      end
-    end
 
     FETCHERS = [DependencyFetcher, IndexFetcher]
